+++ lib/main.dart
@@ -1,11 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+import 'package:sitsmart/utils/routes_name.dart';
 
-import '/core/infoscreen.dart';
-import '/core/homescreen.dart';
+import 'core/info_screen.dart';
+import 'core/home_screen.dart';
 
 void main() {
-  runApp(SitSmart());
+  runApp(
+    const SitSmart(),
+  );
 }
 
 class SitSmart extends StatelessWidget {
@@ -36,9 +39,8 @@ class SitSmart extends StatelessWidget {
         ),
       ),
       getPages: [
-        GetPage(name: HomeScreen.routeName, page: () => HomeScreen()),
-        GetPage(
-            name: InformationScreen.routeName, page: () => InformationScreen()),
+        GetPage(name: RoutesName.root, page: () => HomeScreen()),
+        GetPage(name: RoutesName.sitsmartInfo, page: () => InformationScreen()),
       ],
       home: HomeScreen(),
     );
