+++ lib/utils/routes_name.dart
@@ -1,0 +1,4 @@
+class RoutesName {
+  static const sitsmartInfo = "/sitsmart-info";
+  static const root = '/';
+}
